Updating with latest minting factory
