+++ subgraph.yaml
@@ -418,9 +418,9 @@ dataSources:
           handler: handleAdminUpdateArtistAccessMerkleRootIpfsHash
       file: ./src/mappings/v3/platform-config-mapping.ts
 
-  ## MintingFactory for KODA V3
+  ## MintingFactory V1 for KODA V3
   - kind: ethereum/contract
-    name: MintingFactoryConfig
+    name: MintingFactoryV1
     network: rinkeby
     source:
       address: '0x35668a2Fd45452231a2639edcb887cD45951e597'
@@ -431,6 +431,7 @@ dataSources:
       apiVersion: 0.0.4
       language: wasm/assemblyscript
       entities:
+        - EditionMintedAndListed
         - MintingFactoryCreated
         - AdminMintingPeriodChanged
         - AdminMaxMintsInPeriodChanged
@@ -443,6 +444,8 @@ dataSources:
         - name: KnownOriginV2
           file: ./abis/kodav2/KnownOriginV2.json
       eventHandlers:
+        - event: EditionMintedAndListed(indexed uint256,uint8)
+          handler: handleEditionMintedAndListed
         - event: MintingFactoryCreated()
           handler: handleMintingFactoryCreated
         - event: AdminMintingPeriodChanged(uint256)
@@ -451,11 +454,11 @@ dataSources:
           handler: handleAdminMaxMintsInPeriodChanged
         - event: AdminFrequencyOverrideChanged(address,bool)
           handler: handleAdminFrequencyOverrideChanged
-      file: ./src/mappings/v3/platform-config-mapping.ts
+      file: ./src/mappings/v3/minting-factory-v1-mapping.ts
 
-  ## MintingFactory for KODA V3
+  ## MintingFactory V2 for KODA V3
   - kind: ethereum/contract
-    name: MintingFactoryCreations
+    name: MintingFactoryV2
     network: rinkeby
     source:
       address: '0x35668a2Fd45452231a2639edcb887cD45951e597'
@@ -467,9 +470,13 @@ dataSources:
       language: wasm/assemblyscript
       entities:
         - EditionMintedAndListed
+        - MintingFactoryCreated
+        - AdminMintingPeriodChanged
+        - AdminMaxMintsInPeriodChanged
+        - AdminFrequencyOverrideChanged
       abis:
         - name: MintingFactory
-          file: ./abis/kodav3/MintingFactory.json
+          file: ./abis/kodav3/MintingFactoryV2.json
         - name: KnownOriginV3
           file: ./abis/kodav3/KnownOriginV3.json
         - name: KnownOriginV2
@@ -477,7 +484,17 @@ dataSources:
       eventHandlers:
         - event: EditionMintedAndListed(indexed uint256,uint8)
           handler: handleEditionMintedAndListed
-      file: ./src/mappings/v3/minting-factory-mapping.ts
+        - event: EditionMinted(indexed uint256)
+          handler: handleEditionMinted
+        - event: MintingFactoryCreated()
+          handler: handleMintingFactoryCreated
+        - event: AdminMintingPeriodChanged(uint256)
+          handler: handleAdminMintingPeriodChanged
+        - event: AdminMaxMintsInPeriodChanged(uint256)
+          handler: handleAdminMaxMintsInPeriodChanged
+        - event: AdminFrequencyOverrideChanged(address,bool)
+          handler: handleAdminFrequencyOverrideChanged
+      file: ./src/mappings/v3/minting-factory-v2-mapping.ts
 
   # Token Primary Marketplace V3 (V3 NFT)
   - kind: ethereum/contract
@@ -711,6 +728,8 @@ dataSources:
         - name: KODAV3UpgradableGatedMarketplace
           file: ./abis/kodav3/KODAV3UpgradableGatedMarketplace.json
       eventHandlers:
+        - event: SaleCreated(indexed uint256)
+          handler: handleSaleCreated
         - event: SaleWithPhaseCreated(indexed uint256)
           handler: handleSaleWithPhaseCreated
         - event: PhaseCreated(indexed uint256,indexed uint256)
@@ -725,9 +744,7 @@ dataSources:
           handler: handleSaleResumed
         - event: AdminSetKoCommissionOverrideForSale(indexed uint256,uint256)
           handler: handleAdminUpdatePlatformPrimarySaleCommissionGatedSale
-
           # TODO make config toggled on/off
           # TODO contract Pausable
-          # TODO update PlatformConfig entity
 
       file: ./src/mappings/v3/known-origin-v3-gated-marketplace-mapping.ts
